requirements.txt
pytest~=8.3.3
requests~=2.32.2
pydantic~=2.9.2
allure-pytest~=2.13.5
python-dotenv==1.0.1
faker~=33.1.0